refactor: move all-new Gua files to ElementX/Sources/Gua/ to avoid upstream merge conflicts
All Swift files that have no upstream equivalent are now grouped under
ElementX/Sources/Gua/ so they are clearly isolated from upstream code.
XcodeGen picks them up automatically (path: ../Sources in target.yml).

Moved to Gua/: AppHooks+Gua, PinBubbleField, IdentityServiceClient,
PhoneEntryScreen, OtpEntryScreen, PinSetupScreen, PinChallengeScreen,
ProfileSetupScreen, TwoStepVerificationScreen, PinSetupReminderBanner

New: Gua/Extensions/SettingsFlowCoordinator+Gua.swift
  — presentTwoStepVerification() extracted out of upstream file

Upstream files updated with // GUA FORK: markers for easy diff scanning:
AppRoutes, AuthenticationFlowCoordinator, AuthenticationService,
SettingsFlowCoordinator, SettingsScreenModels, SettingsScreen

Gua.xcodeproj regenerated via xcodegen.
