+++ ElementX/Sources/Application/Navigation/AppRoutes.swift
@@ -44,7 +44,7 @@ enum AppRoute: Hashable {
     case genericCallLink(url: URL)
     /// The settings screen.
     case settings
-    /// The settings screen with the Two-step verification PIN screen pushed on top.
+    /// GUA FORK: The settings screen with the Two-step verification PIN screen pushed on top.
     case settingsTwoStepVerification
     /// The setting screen for key backup.
     case chatBackupSettings
