+++ ElementX/Sources/FlowCoordinators/AuthenticationFlowCoordinator.swift
@@ -23,13 +23,14 @@ class AuthenticationFlowCoordinator: FlowCoordinatorProtocol {
     private let appSettings: AppSettings
     private let analytics: AnalyticsService
     private let userIndicatorController: UserIndicatorControllerProtocol
-    private let identityServiceClient: IdentityServiceClientProtocol?
-    private let usesPhoneLoginHint: Bool
+    private let identityServiceClient: IdentityServiceClientProtocol? // GUA FORK
+    private let usesPhoneLoginHint: Bool // GUA FORK
     
     enum State: StateType {
         /// The state machine hasn't started.
         case initial
         
+        // GUA FORK BEGIN: Gua phone-OTP-PIN onboarding states
         /// The Gua phone-number entry screen (default entry point for normal users).
         case phoneEntryScreen
         /// The Gua OTP entry screen.
@@ -40,6 +41,7 @@ class AuthenticationFlowCoordinator: FlowCoordinatorProtocol {
         case profileSetupScreen
         /// Optional PIN setup step offered to brand-new users after profile setup.
         case pinSetupScreen
+        // GUA FORK END
         
         /// The initial screen shown when you first launch the app.
         case startScreen
@@ -66,12 +68,9 @@ class AuthenticationFlowCoordinator: FlowCoordinatorProtocol {
     enum Event: EventType {
         /// The legacy flow is being started.
         case start
+        // GUA FORK BEGIN: Gua phone-OTP-PIN events
         /// The Gua phone-OTP flow is being started.
         case startPhoneAuth
-        
-        /// Modify the flow using the provisioning parameters in the `userInfo`.
-        case applyProvisioningParameters
-        
         /// The user submitted a phone number (carried via `userInfo`).
         case continueWithPhone
         /// The user dropped into the legacy auth flow from the phone screen.
@@ -79,21 +78,18 @@ class AuthenticationFlowCoordinator: FlowCoordinatorProtocol {
         /// The user cancelled OTP entry to edit their phone number.
         case cancelledOTPEntry
         /// OTP verified for a returning user with two-step verification — prompt for PIN.
-        /// `userInfo` is `PinChallengeContext`.
         case needsPinChallenge
         /// The user backed out of the PIN challenge (returns to OTP entry).
         case cancelledPinChallenge
-        /// OTP verified for a brand-new phone — advance to profile setup. `userInfo` is `ProfileSetupContext`.
+        /// OTP verified for a brand-new phone — advance to profile setup.
         case needsProfileSetup
         /// The user backed out of profile setup (returns to OTP entry).
         case cancelledProfileSetup
         /// Profile setup completed; offer to create a PIN before finishing sign-in.
-        /// `userInfo` is `PendingSignupContext`.
         case offerPinSetup
-        /// Backend rejected the chosen username during signup completion — return to profile setup
-        /// so the user can pick a different one. `userInfo` is the original `ProfileSetupContext`
-        /// (signup token is still valid because the backend only consumes it on success).
+        /// Backend rejected the chosen username during signup completion.
         case usernameTakenDuringSignup
+        // GUA FORK END
         
         /// The user would like to login with a QR code.
         case loginWithQR
