+++ ElementX/Sources/Services/Authentication/AuthenticationService.swift
@@ -92,7 +92,7 @@ class AuthenticationService: AuthenticationServiceProtocol {
         do {
             // The create prompt is broken: https://github.com/element-hq/matrix-authentication-service/issues/3429
             // let prompt: OidcPrompt = flow == .register ? .create : .consent
-            // Don't force the consent screen — Gua is a first-party client and the
+            // GUA FORK: Don't force the consent screen — Gua is a first-party client and the
             // identity-service already gathers/confirms everything. Passing no prompt
             // lets MAS skip consent once a grant exists (frictionless re-login).
             let oidcData = try await client.urlForOauth(oauthConfiguration: appSettings.oidcConfiguration.rustValue,
